refactor: restructure SimpleParticipantFilters into modular components for improved maintainability and reusability
- Created a new folder structure for SimpleParticipantFilters with focused components.
- Implemented ActiveFilterBadges, MoreFiltersPopover, QuickFilters, SearchFilter, and section components.
- Enhanced state management using Jotai for filter states.
- Added utility functions for dynamic filter generation and type definitions.
- Improved performance with memoization and lazy loading of skills.
- Ensured backward compatibility with existing implementations.
