+++ src/features/participants/components/filters/simple-filters/index.ts
@@ -1,0 +1,1 @@
+export type { SimpleParticipantFiltersProps } from "./types";
