+++ src/features/participants/components/filters/simple-participant-filters.tsx
@@ -1,865 +1,3 @@
-"use client";
-
-import { useState, useMemo, useEffect } from "react";
-import { useAtom, useAtomValue, useSetAtom } from "jotai";
-import { Search, X, Filter, ChevronDown } from "lucide-react";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
-import { Combobox, type ComboboxOption } from "@/components/ui/combobox";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-  participantFiltersAtom,
-  participantSearchAtom,
-  hasActiveFiltersAtom,
-  updateFilterAtom,
-  clearFiltersAtom,
-} from "../../atoms/participants-atoms";
-import {
-  type ParticipantFilters as ParticipantFiltersType,
-  type Participant,
-} from "../../types/types";
-import {
-  getUniqueSkills,
-  type SkillsOptions,
-} from "../../actions/get-unique-skills";
-
-interface SimpleParticipantFiltersProps {
-  projects: Array<{ id: string; name: string; acronym: string }>;
-  organizations: Array<{ id: string; name: string; acronym: string }>;
-  districts?: Array<{ id: string; name: string }>;
-  subCounties?: Array<{ id: string; name: string }>;
-  participants?: Participant[]; // Add participants data
-}
-
-// Function to generate dynamic filter options from actual data
-function generateDynamicFilterOptions(participants: Participant[] = []) {
-  // Helper function to create options from participant data
-  const createDynamicOptions = (
-    key: keyof Participant,
-    label: string,
-    transform?: (value: string) => string
-  ) => {
-    const uniqueValues = participants
-      .map(p => p[key])
-      .filter(
-        (value, index, arr) =>
-          value &&
-          value !== "all" &&
-          typeof value === "string" &&
-          arr.indexOf(value) === index
-      )
-      .sort();
-
-    return {
-      key,
-      label,
-      values: [
-        { value: "all", label: "All" },
-        ...uniqueValues.map(value => ({
-          value: value as string,
-          label: transform
-            ? transform(value as string)
-            : capitalizeFirst(value as string),
-        })),
-      ],
-    };
-  };
-
-  // Helper function to capitalize first letter
-  const capitalizeFirst = (str: string) =>
-    str.charAt(0).toUpperCase() + str.slice(1).replace(/[-_]/g, " ");
-
-  // Core demographic filters
-  const sexOptions = createDynamicOptions("sex", "Gender", value =>
-    value === "male"
-      ? "Male"
-      : value === "female"
-        ? "Female"
-        : capitalizeFirst(value)
-  );
-
-  const hasAgeData = participants.some(p => p.age || p.dateOfBirth);
-  const ageGroupOptions = hasAgeData
-    ? {
-        key: "ageGroup",
-        label: "Age",
-        values: [
-          { value: "all", label: "All Ages" },
-          { value: "young", label: "Youth (15-35)" },
-          { value: "adult", label: "Adults (36-59)" },
-          { value: "older", label: "Elderly (60+)" },
-        ],
-      }
-    : null;
-
-  // Employment and enterprise filters
-  const employmentTypeOptions = createDynamicOptions(
-    "employmentType",
-    "Employment Type"
-  );
-
-  // Add fallback options for employment type if no data exists
-  if (employmentTypeOptions.values.length <= 1) {
-    employmentTypeOptions.values = [
-      { value: "all", label: "All" },
-      { value: "formal", label: "Formal Employment" },
-      { value: "informal", label: "Informal Employment" },
-      { value: "self-employed", label: "Self-Employed" },
-      { value: "unemployed", label: "Unemployed" },
-    ];
-  }
-
-  const employmentSectorOptions = createDynamicOptions(
-    "employmentSector",
-    "Employment Sector"
-  );
-  const enterpriseSectorOptions = createDynamicOptions(
-    "enterpriseSector",
-    "Enterprise Sector"
-  );
-  const businessScaleOptions = createDynamicOptions(
-    "businessScale",
-    "Business Scale"
-  );
-
-  // Skills filters
-  const vocationalSkillsOptions = {
-    key: "hasVocationalSkills",
-    label: "Vocational Skills",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Has Skills" },
-      { value: "no", label: "No Skills" },
-    ],
-  };
-
-  const softSkillsOptions = {
-    key: "hasSoftSkills",
-    label: "Soft Skills",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Has Skills" },
-      { value: "no", label: "No Skills" },
-    ],
-  };
-
-  const businessSkillsOptions = {
-    key: "hasBusinessSkills",
-    label: "Business Skills",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Has Skills" },
-      { value: "no", label: "No Skills" },
-    ],
-  };
-
-  // Education and demographics
-  const educationLevelOptions = createDynamicOptions(
-    "educationLevel",
-    "Education"
-  );
-  const maritalStatusOptions = createDynamicOptions(
-    "maritalStatus",
-    "Marital Status"
-  );
-  const sourceOfIncomeOptions = createDynamicOptions(
-    "sourceOfIncome",
-    "Income Source"
-  );
-  const populationSegmentOptions = createDynamicOptions(
-    "populationSegment",
-    "Population Segment"
-  );
-
-  // VSLA and enterprise ownership
-  const vslaOptions = {
-    key: "isSubscribedToVSLA",
-    label: "VSLA",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Member" },
-      { value: "no", label: "Non-Member" },
-    ],
-  };
-
-  const enterpriseOwnershipOptions = {
-    key: "ownsEnterprise",
-    label: "Enterprise Owner",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Owns Enterprise" },
-      { value: "no", label: "No Enterprise" },
-    ],
-  };
-
-  // Special demographic filters
-  const pwdOptions = {
-    key: "isPWD",
-    label: "PWD",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "PWD" },
-      { value: "no", label: "Non-PWD" },
-    ],
-  };
-
-  const studentOptions = {
-    key: "isActiveStudent",
-    label: "Student",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Student" },
-      { value: "no", label: "Non-Student" },
-    ],
-  };
-
-  const teenMotherOptions = {
-    key: "isTeenMother",
-    label: "Teen Mother",
-    values: [
-      { value: "all", label: "All" },
-      { value: "yes", label: "Teen Mother" },
-      { value: "no", label: "Not Teen Mother" },
-    ],
-  };
-
-  // Return organized filter groups
-  return {
-    quick: [sexOptions, ageGroupOptions, employmentTypeOptions, vslaOptions]
-      .filter(Boolean)
-      .filter((item): item is NonNullable<typeof item> => item !== null),
-
-    enterprise: [
-      enterpriseOwnershipOptions,
-      enterpriseSectorOptions,
-      businessScaleOptions,
-      sourceOfIncomeOptions,
-    ].filter(option => option.values.length > 1),
-
-    skills: [
-      vocationalSkillsOptions,
-      softSkillsOptions,
-      businessSkillsOptions,
-      studentOptions,
-    ],
-
-    demographics: [
-      educationLevelOptions,
-      maritalStatusOptions,
-      populationSegmentOptions,
-      pwdOptions,
-      teenMotherOptions,
-    ].filter(option => option.values.length > 1),
-
-    employment: [employmentTypeOptions, employmentSectorOptions].filter(
-      option => option.values.length > 1
-    ),
-  };
-}
-
-export function SimpleParticipantFilters({
-  projects,
-  organizations,
-  districts = [],
-  subCounties = [],
-  participants = [],
-}: SimpleParticipantFiltersProps) {
-  const [showAdvanced, setShowAdvanced] = useState(false);
-  const [skillsOptions, setSkillsOptions] = useState<SkillsOptions>({
-    vocationalSkills: [],
-    softSkills: [],
-    businessSkills: [],
-  });
-  const [isLoadingSkills, setIsLoadingSkills] = useState(true);
-  const [localSearchValue, setLocalSearchValue] = useState("");
-
-  // Jotai state
-  const filters = useAtomValue(participantFiltersAtom);
-  const [searchValue, setSearchValue] = useAtom(participantSearchAtom);
-  const hasActiveFilters = useAtomValue(hasActiveFiltersAtom);
-  const updateFilter = useSetAtom(updateFilterAtom);
-  const clearFilters = useSetAtom(clearFiltersAtom);
-
-  // Initialize local search value from Jotai state
-  useEffect(() => {
-    setLocalSearchValue(searchValue);
-  }, [searchValue]);
-
-  // Debounced search effect
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setSearchValue(localSearchValue);
-    }, 300); // 300ms debounce
-
-    return () => clearTimeout(timer);
-  }, [localSearchValue, setSearchValue]);
-
-  // Load unique skills on component mount
-  useEffect(() => {
-    const loadSkills = async () => {
-      setIsLoadingSkills(true);
-      try {
-        const skills = await getUniqueSkills();
-        setSkillsOptions(skills);
-      } catch (error) {
-        console.error("Failed to load skills:", error);
-      } finally {
-        setIsLoadingSkills(false);
-      }
-    };
-
-    loadSkills();
-  }, []);
-
-  // Generate organized filter options based on actual data
-  const filterGroups = useMemo(
-    () => generateDynamicFilterOptions(participants),
-    [participants]
-  );
-
-  // Helper function to convert skills array to ComboboxOption format
-  const convertSkillsToOptions = (skills: string[]): ComboboxOption[] => {
-    return [
-      { value: "all", label: "All Skills" },
-      ...skills.map(skill => ({
-        value: skill,
-        label: skill,
-      })),
-    ];
-  };
-
-  const handleFilterUpdate = (
-    key: keyof ParticipantFiltersType,
-    value: string
-  ) => {
-    updateFilter({ key, value });
-  };
-
-  // Helper function to safely access filter values
-  const getFilterValue = (key: string): string => {
-    return (filters as Record<string, string>)[key] || "all";
-  };
-
-  const handleClearFilters = () => {
-    clearFilters();
-  };
-
-  const getActiveFiltersCount = () => {
-    return Object.entries(filters).filter(([key, value]) => {
-      if (key === "search") return false;
-      return value !== "" && value !== "all";
-    }).length;
-  };
-
-  const getActiveFilterBadges = () => {
-    const allFilters = [
-      ...filterGroups.quick,
-      ...filterGroups.enterprise,
-      ...filterGroups.skills,
-      ...filterGroups.demographics,
-      ...filterGroups.employment,
-    ];
-
-    return Object.entries(filters)
-      .filter(
-        ([key, value]) => key !== "search" && value !== "" && value !== "all"
-      )
-      .map(([key, value]) => {
-        const filterOption = allFilters.find(f => f?.key === key);
-        const displayValue = filterOption
-          ? filterOption.values.find(v => v.value === value)?.label || value
-          : value;
-
-        return (
-          <Badge key={key} variant="secondary" className="gap-1 text-xs">
-            {filterOption?.label || key}: {displayValue}
-            <X
-              className="hover:text-destructive h-3 w-3 cursor-pointer"
-              onClick={() =>
-                handleFilterUpdate(key as keyof ParticipantFiltersType, "all")
-              }
-            />
-          </Badge>
-        );
-      });
-  };
-
-  const activeFiltersCount = getActiveFiltersCount();
-
-  return (
-    <div className="space-y-3">
-      {/* Main Filter Bar */}
-      <div className="flex flex-wrap items-center gap-3">
-        {/* Search */}
-        <div className="relative min-w-[250px] flex-1">
-          <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-            Search Participants
-          </label>
-          <div className="relative">
-            <Search className="text-muted-foreground absolute top-1/2 left-3 h-4 w-4 -translate-y-1/2" />
-            <Input
-              placeholder="Search participants..."
-              value={localSearchValue}
-              onChange={e => setLocalSearchValue(e.target.value)}
-              className="pl-9"
-            />
-          </div>
-        </div>
-
-        {/* Quick Filters */}
-        {filterGroups.quick.map(filter => (
-          <div key={filter.key} className="min-w-[140px]">
-            <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-              {filter.label}
-            </label>
-            <Select
-              value={getFilterValue(filter.key)}
-              onValueChange={value =>
-                handleFilterUpdate(
-                  filter.key as keyof ParticipantFiltersType,
-                  value
-                )
-              }
-            >
-              <SelectTrigger className="w-[140px]">
-                <SelectValue placeholder={filter.label} />
-              </SelectTrigger>
-              <SelectContent>
-                {filter.values.map(option => (
-                  <SelectItem key={option.value} value={option.value}>
-                    {option.label}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        ))}
-
-        {/* More Filters Popover */}
-        <div className="min-w-[140px]">
-          <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-            Additional Filters
-          </label>
-          <Popover open={showAdvanced} onOpenChange={setShowAdvanced}>
-            <PopoverTrigger asChild>
-              <Button
-                variant="outline"
-                size="default"
-                className="w-[140px] gap-2"
-              >
-                <Filter className="h-4 w-4" />
-                More
-                {activeFiltersCount > 4 && (
-                  <Badge
-                    variant="secondary"
-                    className="ml-1 h-5 w-5 p-0 text-xs"
-                  >
-                    +{activeFiltersCount - 4}
-                  </Badge>
-                )}
-                <ChevronDown
-                  className={`h-4 w-4 transition-transform duration-200 ${showAdvanced ? "rotate-180" : ""}`}
-                />
-              </Button>
-            </PopoverTrigger>
-            <PopoverContent
-              align="end"
-              className="animate-in data-[state=open]:fade-in-0 data-[state=open]:zoom-in-95 data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 h-[500px] w-[85vw] max-w-5xl overflow-y-auto"
-            >
-              <div className="space-y-4">
-                <div className="flex items-center justify-between">
-                  <h4 className="font-semibold">Additional Filters</h4>
-                </div>
-
-                {/* Organization & Location Section */}
-                <div className="space-y-3">
-                  <h5 className="text-muted-foreground border-b pb-1 text-sm font-medium">
-                    Organization & Location
-                  </h5>
-                  <div className="grid grid-cols-2 gap-3 lg:grid-cols-3 xl:grid-cols-4">
-                    <div className="space-y-2">
-                      <label className="text-xs font-medium">Project</label>
-                      <Select
-                        value={filters.project}
-                        onValueChange={value =>
-                          handleFilterUpdate("project", value)
-                        }
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select project..." />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="all">All Projects</SelectItem>
-                          {projects.map(project => (
-                            <SelectItem key={project.id} value={project.id}>
-                              {project.acronym}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-
-                    <div className="space-y-1">
-                      <label className="text-xs font-medium">
-                        Organization
-                      </label>
-                      <Select
-                        value={filters.organization}
-                        onValueChange={value =>
-                          handleFilterUpdate("organization", value)
-                        }
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select organization..." />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="all">All Organizations</SelectItem>
-                          {organizations.map(org => (
-                            <SelectItem key={org.id} value={org.id}>
-                              {org.acronym}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-
-                    <div className="space-y-1">
-                      <label className="text-xs font-medium">District</label>
-                      <Select
-                        value={filters.district}
-                        onValueChange={value =>
-                          handleFilterUpdate("district", value)
-                        }
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select district..." />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="all">All Districts</SelectItem>
-                          {districts.map(district => (
-                            <SelectItem key={district.id} value={district.id}>
-                              {district.name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-
-                    <div className="space-y-1">
-                      <label className="text-xs font-medium">Sub County</label>
-                      <Select
-                        value={filters.subCounty}
-                        onValueChange={value =>
-                          handleFilterUpdate("subCounty", value)
-                        }
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select sub county..." />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="all">All Sub Counties</SelectItem>
-                          {subCounties.map(subCounty => (
-                            <SelectItem key={subCounty.id} value={subCounty.id}>
-                              {subCounty.name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Enterprise & Business Section */}
-                {filterGroups.enterprise.length > 0 && (
-                  <div className="space-y-3">
-                    <h5 className="text-muted-foreground border-b pb-1 text-sm font-medium">
-                      Enterprise & Business
-                    </h5>
-                    <div className="grid grid-cols-2 gap-3 lg:grid-cols-3 xl:grid-cols-4">
-                      {filterGroups.enterprise.map(filter => (
-                        <div key={filter.key} className="space-y-1">
-                          <label className="text-xs font-medium">
-                            {filter.label}
-                          </label>
-                          <Select
-                            value={getFilterValue(filter.key)}
-                            onValueChange={value =>
-                              handleFilterUpdate(
-                                filter.key as keyof ParticipantFiltersType,
-                                value
-                              )
-                            }
-                          >
-                            <SelectTrigger>
-                              <SelectValue
-                                placeholder={`Select ${filter.label.toLowerCase()}...`}
-                              />
-                            </SelectTrigger>
-                            <SelectContent>
-                              {filter.values.map(option => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                {/* Skills & Education Section */}
-                <div className="space-y-4">
-                  <h5 className="text-muted-foreground border-b pb-1 text-sm font-medium">
-                    Skills & Education
-                  </h5>
-
-                  {/* General Skills Status */}
-                  <div>
-                    <h6 className="mb-2 text-xs font-medium tracking-wide text-gray-600 uppercase">
-                      General Skills Status
-                    </h6>
-                    <div className="grid grid-cols-2 gap-3 lg:grid-cols-3 xl:grid-cols-4">
-                      {filterGroups.skills.map(filter => (
-                        <div key={filter.key} className="space-y-1">
-                          <label className="text-xs font-medium">
-                            {filter.label}
-                          </label>
-                          <Select
-                            value={getFilterValue(filter.key)}
-                            onValueChange={value =>
-                              handleFilterUpdate(
-                                filter.key as keyof ParticipantFiltersType,
-                                value
-                              )
-                            }
-                          >
-                            <SelectTrigger className="h-9">
-                              <SelectValue
-                                placeholder={`Select ${filter.label.toLowerCase()}...`}
-                              />
-                            </SelectTrigger>
-                            <SelectContent>
-                              {filter.values.map(option => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-
-                  {/* Specific Skills Filters */}
-                  <div>
-                    <h6 className="mb-3 text-xs font-medium tracking-wide text-gray-600 uppercase">
-                      Specific Skills
-                    </h6>
-                    <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-                      {/* Specific Vocational Skills */}
-                      <div className="space-y-2">
-                        <label className="text-xs font-medium text-gray-700">
-                          Vocational Skill
-                        </label>
-                        <Combobox
-                          options={convertSkillsToOptions(
-                            skillsOptions.vocationalSkills
-                          )}
-                          value={filters.specificVocationalSkill || "all"}
-                          onValueChange={value =>
-                            handleFilterUpdate("specificVocationalSkill", value)
-                          }
-                          placeholder={
-                            isLoadingSkills
-                              ? "Loading skills..."
-                              : "Select vocational skill..."
-                          }
-                          emptyMessage="No vocational skills found"
-                          disabled={isLoadingSkills}
-                          className="h-9"
-                        />
-                      </div>
-
-                      {/* Specific Soft Skills */}
-                      <div className="space-y-2">
-                        <label className="text-xs font-medium text-gray-700">
-                          Soft Skill
-                        </label>
-                        <Combobox
-                          options={convertSkillsToOptions(
-                            skillsOptions.softSkills
-                          )}
-                          value={filters.specificSoftSkill || "all"}
-                          onValueChange={value =>
-                            handleFilterUpdate("specificSoftSkill", value)
-                          }
-                          placeholder={
-                            isLoadingSkills
-                              ? "Loading skills..."
-                              : "Select soft skill..."
-                          }
-                          emptyMessage="No soft skills found"
-                          disabled={isLoadingSkills}
-                          className="h-9"
-                        />
-                      </div>
-
-                      {/* Business Skills - Only show if there are business skills available */}
-                      {skillsOptions.businessSkills.length > 0 && (
-                        <div className="space-y-2">
-                          <label className="text-xs font-medium text-gray-700">
-                            Business Skill
-                          </label>
-                          <Combobox
-                            options={convertSkillsToOptions(
-                              skillsOptions.businessSkills
-                            )}
-                            value={filters.specificBusinessSkill || "all"}
-                            onValueChange={value =>
-                              handleFilterUpdate("specificBusinessSkill", value)
-                            }
-                            placeholder={
-                              isLoadingSkills
-                                ? "Loading skills..."
-                                : "Select business skill..."
-                            }
-                            emptyMessage="No business skills found"
-                            disabled={isLoadingSkills}
-                            className="h-9"
-                          />
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                </div>
-
-                {/* Demographics & Personal Section */}
-                {filterGroups.demographics.length > 0 && (
-                  <div className="space-y-3">
-                    <h5 className="text-muted-foreground border-b pb-1 text-sm font-medium">
-                      Demographics & Personal
-                    </h5>
-                    <div className="grid grid-cols-4 gap-3">
-                      {filterGroups.demographics.map(filter => (
-                        <div key={filter.key} className="space-y-1">
-                          <label className="text-xs font-medium">
-                            {filter.label}
-                          </label>
-                          <Select
-                            value={getFilterValue(filter.key)}
-                            onValueChange={value =>
-                              handleFilterUpdate(
-                                filter.key as keyof ParticipantFiltersType,
-                                value
-                              )
-                            }
-                          >
-                            <SelectTrigger>
-                              <SelectValue
-                                placeholder={`Select ${filter.label.toLowerCase()}...`}
-                              />
-                            </SelectTrigger>
-                            <SelectContent>
-                              {filter.values.map(option => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                {/* Employment Section */}
-                {filterGroups.employment.length > 0 && (
-                  <div className="space-y-3">
-                    <h5 className="text-muted-foreground border-b pb-1 text-sm font-medium">
-                      Employment
-                    </h5>
-                    <div className="grid grid-cols-4 gap-3">
-                      {filterGroups.employment.map(filter => (
-                        <div key={filter.key} className="space-y-1">
-                          <label className="text-xs font-medium">
-                            {filter.label}
-                          </label>
-                          <Select
-                            value={getFilterValue(filter.key)}
-                            onValueChange={value =>
-                              handleFilterUpdate(
-                                filter.key as keyof ParticipantFiltersType,
-                                value
-                              )
-                            }
-                          >
-                            <SelectTrigger>
-                              <SelectValue
-                                placeholder={`Select ${filter.label.toLowerCase()}...`}
-                              />
-                            </SelectTrigger>
-                            <SelectContent>
-                              {filter.values.map(option => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-              </div>
-            </PopoverContent>
-          </Popover>
-        </div>
-      </div>
-
-      {/* Active Filter Badges */}
-      {hasActiveFilters && (
-        <div className="animate-in fade-in-0 slide-in-from-top-1 flex flex-wrap items-center gap-2 duration-300">
-          <span className="text-muted-foreground text-sm">Active filters:</span>
-          {getActiveFilterBadges()}
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={handleClearFilters}
-            className="ml-auto h-6 px-2 text-xs"
-          >
-            Clear All
-          </Button>
-        </div>
-      )}
-    </div>
-  );
-}
+// This file has been split into smaller components in the simple-filters/ folder
+// Re-export the main component for backward compatibility
+export type { SimpleParticipantFiltersProps } from "./simple-filters/types";
